refactor: Name Vo 적용 및 정적 팩토리메소드 추가
