+++ src/main/java/racingcar/domain/Name.java
@@ -1,0 +1,39 @@
+package racingcar.domain;
+
+import java.util.Objects;
+
+public class Name {
+
+    private static final int NAME_MAX_LENGTH = 5;
+
+    private final String name;
+
+    public Name(String name) {
+        checkNameNull(name);
+        checkNameEmpty(name);
+        checkNameLength(name);
+        this.name = name;
+    }
+
+    private void checkNameNull(String name) {
+        if (Objects.isNull(name)) {
+            throw new IllegalArgumentException("자동차 이름은 null이 들어올 수 없다.");
+        }
+    }
+
+    private void checkNameEmpty(String name) {
+        if (name.equals("")) {
+            throw new IllegalArgumentException("자동차 이름은 공백이 들어올 수 없다.");
+        }
+    }
+
+    private void checkNameLength(String name) {
+        if (name.length() > NAME_MAX_LENGTH) {
+            throw new IllegalArgumentException("자동차 이름은 5자를 초과할 수 없다.");
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
