+++ src/main/java/racingcar/domain/RacingCar.java
@@ -1,43 +1,19 @@
 package racingcar.domain;
 
-import java.util.Objects;
-
 public class RacingCar {
 
-    private static final int NAME_MAX_LENGTH = 5;
     private static final int DEFAULT_POSITION = 0;
 
     private final RandomMoveGenerator generator = new RandomMoveGenerator();
-    private final String name;
+    private final Name name;
     private int position = DEFAULT_POSITION;
 
-    public RacingCar(String name) {
-        checkNameNull(name);
-        checkNameEmpty(name);
-        checkNameLength(name);
+    public RacingCar(Name name) {
         this.name = name;
     }
 
-    private void checkNameNull(String name) {
-        if (Objects.isNull(name)) {
-            throw new IllegalArgumentException("자동차 이름은 null이 들어올 수 없다.");
-        }
-    }
-
-    private void checkNameEmpty(String name) {
-        if (name.equals("")) {
-            throw new IllegalArgumentException("자동차 이름은 공백이 들어올 수 없다.");
-        }
-    }
-
-    private void checkNameLength(String name) {
-        if (name.length() > NAME_MAX_LENGTH) {
-            throw new IllegalArgumentException("자동차 이름은 5자를 초과할 수 없다.");
-        }
-    }
-
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     public int getPosition() {
