+++ src/test/java/racingcar/domain/RacingCarTest.java
@@ -1,7 +1,6 @@
 package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -14,31 +13,8 @@ public class RacingCarTest {
 
     @BeforeEach
     void setUp() {
-        car = new RacingCar("name");
-    }
-
-    @Test
-    @DisplayName("자동차 이름이 null이 들어올 경우 예외가 발생해야 한다.")
-    void checkRacingCarNull() {
-        assertThatExceptionOfType(IllegalArgumentException.class)
-            .isThrownBy(() -> new RacingCar(null))
-            .withMessageMatching("자동차 이름은 null이 들어올 수 없다.");
-    }
-
-    @Test
-    @DisplayName("자동차의 이름이 공백이 들어올 경우 예외가 발생해야 한다.")
-    void checkRacingCarEmpty() {
-        assertThatExceptionOfType(IllegalArgumentException.class)
-            .isThrownBy(() -> new RacingCar(""))
-            .withMessageMatching("자동차 이름은 공백이 들어올 수 없다.");
-    }
-
-    @Test
-    @DisplayName("자동차의 이름이 5자가 초과될 경우 예외가 발생해야 한다.")
-    void checkRacingCarNameLength() {
-        assertThatExceptionOfType(IllegalArgumentException.class)
-            .isThrownBy(() -> new RacingCar("abcede"))
-            .withMessageMatching("자동차 이름은 5자를 초과할 수 없다.");
+        Name name = new Name("name");
+        car = new RacingCar(name);
     }
 
     @Test
