+++ src/test/java/racingcar/domain/RacingCarsTest.java
@@ -11,21 +11,21 @@ class RacingCarsTest {
     @Test
     @DisplayName("자동차 이름을 받아 자동차를 생성한다.")
     void createRacingCars() {
-        RacingCars racingCars = new RacingCars("pobi,crong,honux");
+        RacingCars racingCars = RacingCars.from("pobi,crong,honux");
         assertThat(racingCars.getCarNames()).containsExactly("pobi", "crong", "honux");
     }
 
     @Test
     @DisplayName("자동차 이름의 공백을 제거한다.")
     void trimBlank() {
-        RacingCars racingCars = new RacingCars("pobi , cr ong , honux");
+        RacingCars racingCars = RacingCars.from("pobi , cr ong , honux");
         assertThat(racingCars.getCarNames()).containsExactly("pobi", "crong", "honux");
     }
 
     @Test
     @DisplayName("자동차 우승자 이름을 계산한다.")
     void calculateWinnerNames() {
-        RacingCars racingCars = new RacingCars("pobi,crong,honux");
+        RacingCars racingCars = RacingCars.from("pobi,crong,honux");
         List<String> winnerNames = racingCars.calculateWinnerNames();
         assertThat(winnerNames).containsExactly("pobi", "crong", "honux");
     }
